Add and update index and eggfriedrice html files
This puts into practice commits and git commands. But this also updated
the files for the recipe project. An index.html was created, with a
header. And the
first recipe was created, with descriptive header tags, paragraphs, and
lists were used.
The lists were used to organize and complete the pages structure in
eggfriedrice.html.
Then in index.html within an ul and inside a anchor tag, we link
eggfriedrice.html. This will be done very similarly in the following
iteration for two other recipes.